+++ robot_software/vision/depth_estimation/calibration-images/depth-estimation-1.jpg
