+++ robot_software/vision/depth_estimation/calibration-images/depth-estimation-2.jpg
