+++ robot_software/vision/depth_estimation/calibration-images/depth-estimation-reference.jpg
