+++ robot_software/vision/depth_estimation/calibration-images/scale-factor-k.jpg
