+++ robot_software/vision/depth_estimation/results/test-25.jpg
